feat: move runtime test utils to test dir
