+++ packages/runtime/plugin-runtime/tests/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './initial';
+export * from './render';
+export * from './wrap';
